refactor: useImageUploader을 common 아래로 이동
