+++ frontend/src/hooks/common/index.ts
@@ -4,3 +4,4 @@ export { default as useIntersectionObserver } from './useIntersectionObserver';
 export { default as useRoutePage } from './useRoutePage';
 export { default as useScroll } from './useScroll';
 export { default as useSortOption } from './useSortOption';
+export { default as useImageUploader } from './useImageUploader';
