+++ frontend/src/hooks/common/useImageUploader.ts
